@@ -1,5 +1,11 @@
-import { default as HelloWorld } from './ui/components/HelloWorld';
+import { default as Pulse } from './ui/components/Pulse';
+import { default as Channel } from './ui/components/Channel'
+import { default as Since } from './ui/components/Since'
+import { default as Impulse} from './ui/components/Impulse'
 
 export default {
-  HelloWorld,
+  Pulse,
+  Impulse,
+  Since,
+  Channel
 };
